+++ release_notes.md
@@ -16,3 +16,4 @@
   - This fixes a bug with worker indexing where we are shutting down worker channels and creating a new channel that never
     gets properly initialized as the invocation buffers are not created - this leads to a "Did not find initialized workers" error.
 - Check if a blob container or table exists before trying to create it (#9555)
+- Limit dotnet-isolated specialization to 64 bit host process (#9548)